FIX: compile warning
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@2678 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/groupwise_rmi_warp.cxx
@@ -185,7 +185,7 @@ doMain( int argc, char ** argv )
   functional->SetForceZeroSumFirstN( ForceZeroSumFirstN );
   functional->SetForceZeroSumNoAffine( ForceZeroSumNoAffine );
   if ( UserBackgroundFlag )
-    functional->SetUserBackgroundValue( UserBackgroundValue );
+    functional->SetUserBackgroundValue( static_cast<byte>( UserBackgroundValue ) );
   
   // disable parameters with less than 1% of maximum contribution
   functional->SetPartialGradientMode( (PartialGradientThreshold > 0) , PartialGradientThreshold );
